chore: cleanup workspace, split out util methods & removed unused commented grid layout code
